# crud/src/docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /workCenter/update/{id}:
    put:
      tags:
        - work-center-controller
      operationId: updateWorkCenter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkCenter'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkCenter'
  /employee/update:
    put:
      tags:
        - employee-controller
      operationId: update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Employee'
  /department/update/{id}:
    put:
      tags:
        - department-controller
      operationId: updateDepartment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Department'
  /workCenter/save:
    post:
      tags:
        - work-center-controller
      operationId: saveWorkCenter
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkCenter'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkCenter'
  /employee/save:
    post:
      tags:
        - employee-controller
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Employee'
  /department/save:
    post:
      tags:
        - department-controller
      operationId: saveDepartment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Department'
  /workCenter/list:
    get:
      tags:
        - work-center-controller
      operationId: getAllWorkCenters
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkCenter'
  /workCenter/list/{id}:
    get:
      tags:
        - work-center-controller
      operationId: getWorkCenterById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WorkCenter'
  /employee:
    get:
      tags:
        - employee-controller
      operationId: hello
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /employee/list:
    get:
      tags:
        - employee-controller
      operationId: getAll
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
  /employee/list/{employeeId}:
    get:
      tags:
        - employee-controller
      operationId: getById
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Employee'
  /department/list:
    get:
      tags:
        - department-controller
      operationId: getAllDepartments
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
  /department/list/{id}:
    get:
      tags:
        - department-controller
      operationId: getDepartmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Department'
  /workCenter/delete/{id}:
    delete:
      tags:
        - work-center-controller
      operationId: deleteWorkCenter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /employee/delete/{employeeId}:
    delete:
      tags:
        - employee-controller
      operationId: delete
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /department/delete/{id}:
    delete:
      tags:
        - department-controller
      operationId: deleteDepartment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    WorkCenter:
      type: object
      properties:
        id:
          type: integer
          format: int64
        location:
          type: string
    Department:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Employee:
      required:
        - address
        - firstName
        - lastName
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        email:
          type: string
        workCenter:
          $ref: '#/components/schemas/WorkCenter'
        department:
          $ref: '#/components/schemas/Department'